requirements_my.txt
aiohttp
aiosignal
annotated-types
anyio
attrs
blinker
certifi
charset-normalizer
click
colorama
dataclasses-json
distro
docx2txt
fastapi
filelock
Flask
frozenlist
fsspec
gitdb
GitPython
greenlet
grpcio
grpcio-tools
h11
h2
hpack
httpcore
httpx
huggingface-hub
hyperframe
idna
itsdangerous
Jinja2
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications

langchain
langchain-cli
langchain-community
langchain-core
langchain-experimental
langchain-openai
langchain-text-splitters
langserve
langsmith

libcst
markdown-it-py
MarkupSafe
marshmallow
mdurl
multidict
mypy-extensions
numpy
openai
orjson
packaging
pillow
portalocker
protobuf
pydantic
pydantic_core
Pygments
pypdf
pyproject-toml
python-dotenv
pywin32
PyYAML
qdrant-client
referencing
regex
requests
rich
rpds-py
safetensors
setuptools
shellingham
smmap
sniffio
SQLAlchemy
sse-starlette
starlette
tenacity
tiktoken
tokenizers
toml
tomlkit
tqdm
transformers
typer
typing-inspect
typing_extensions
urllib3
uvicorn
Werkzeug
wheel
yarl

# more requirements
pandas
ipykernel
